+++ src/index.ts
@@ -1,3 +1,7 @@
 export { typeFlag } from './type-flag';
 export { getFlag } from './get-flag';
-export type { TypeFlag, Flags } from './types';
+export type {
+	TypeFlag,
+	TypeFlagOptions,
+	Flags,
+} from './types';
